Refactor test names for clarity and consistency
