+++ suites/metrics/large/conversations.test.ts
@@ -1,11 +1,11 @@
 import { getFixedNames } from "@helpers/client";
 import { logError } from "@helpers/logger";
-import { verifyMessageStream } from "@helpers/streams";
+import { verifyNewConversationStream } from "@helpers/streams";
 import { setupTestLifecycle } from "@helpers/vitest";
 import { getInboxIds } from "@inboxes/utils";
 import { typeofStream } from "@workers/main";
 import { getWorkers, type WorkerManager } from "@workers/manager";
-import type { Group } from "@xmtp/node-sdk";
+import { type Group } from "@xmtp/node-sdk";
 import { afterAll, describe, expect, it } from "vitest";
 import {
   m_large_BATCH_SIZE,
@@ -15,10 +15,11 @@ import {
   type SummaryEntry,
 } from "./helpers";
 
-const testName = "m_large_messages";
+const testName = "m_large_conversations";
 
 describe(testName, async () => {
   let workers: WorkerManager;
+
   let newGroup: Group;
 
   const summaryMap: Record<number, SummaryEntry> = {};
@@ -26,7 +27,7 @@ describe(testName, async () => {
   workers = await getWorkers(
     getFixedNames(m_large_WORKER_COUNT),
     testName,
-    typeofStream.Message,
+    typeofStream.Conversation,
   );
 
   let customDuration: number | undefined = undefined;
@@ -48,30 +49,26 @@ describe(testName, async () => {
     i <= m_large_TOTAL;
     i += m_large_BATCH_SIZE
   ) {
-    it(`should deliver messages to all ${i} group members within acceptable time limits and verify stream consistency`, async () => {
+    it(`conversationStream-${i}: should create ${i}-member group and verify all workers receive new conversation notifications within acceptable time`, async () => {
       try {
         const creator = workers.getCreator();
         newGroup = (await creator.client.conversations.newGroup(
           getInboxIds(i),
         )) as Group;
-        await newGroup.sync();
-        await newGroup.addMembers(
-          workers.getAllButCreator().map((worker) => worker.client.inboxId),
-        );
-        await newGroup.sync();
-        const verifyResult = await verifyMessageStream(
+        // Use the dedicated conversation stream verification helper
+        const verifyResult = await verifyNewConversationStream(
           newGroup,
           workers.getAllButCreator(),
         );
 
+        setCustomDuration(verifyResult.averageEventTiming);
+        expect(verifyResult.allReceived).toBe(true);
+
         // Save metrics
         summaryMap[i] = {
           ...(summaryMap[i] ?? { groupSize: i }),
-          messageStreamTimeMs: verifyResult.averageEventTiming,
+          conversationStreamTimeMs: verifyResult.averageEventTiming,
         };
-
-        setCustomDuration(verifyResult.averageEventTiming);
-        expect(verifyResult.allReceived).toBe(true);
       } catch (e) {
         logError(e, expect.getState().currentTestName);
         throw e;
@@ -79,7 +76,6 @@ describe(testName, async () => {
     });
   }
 
-  // Aft
   // After all tests have run, output a concise summary of all timings per group size
   afterAll(() => {
     saveLog(summaryMap);
