+++ suites/metrics/large/messages.test.ts
@@ -1,11 +1,11 @@
 import { getFixedNames } from "@helpers/client";
 import { logError } from "@helpers/logger";
-import { verifyNewConversationStream } from "@helpers/streams";
+import { verifyMessageStream } from "@helpers/streams";
 import { setupTestLifecycle } from "@helpers/vitest";
 import { getInboxIds } from "@inboxes/utils";
 import { typeofStream } from "@workers/main";
 import { getWorkers, type WorkerManager } from "@workers/manager";
-import { type Group } from "@xmtp/node-sdk";
+import type { Group } from "@xmtp/node-sdk";
 import { afterAll, describe, expect, it } from "vitest";
 import {
   m_large_BATCH_SIZE,
@@ -15,11 +15,10 @@ import {
   type SummaryEntry,
 } from "./helpers";
 
-const testName = "m_large_conversations";
+const testName = "m_large_messages";
 
 describe(testName, async () => {
   let workers: WorkerManager;
-
   let newGroup: Group;
 
   const summaryMap: Record<number, SummaryEntry> = {};
@@ -27,7 +26,7 @@ describe(testName, async () => {
   workers = await getWorkers(
     getFixedNames(m_large_WORKER_COUNT),
     testName,
-    typeofStream.Conversation,
+    typeofStream.Message,
   );
 
   let customDuration: number | undefined = undefined;
@@ -49,26 +48,30 @@ describe(testName, async () => {
     i <= m_large_TOTAL;
     i += m_large_BATCH_SIZE
   ) {
-    it(`should create ${i}-member group and verify all workers receive new conversation notifications within acceptable time`, async () => {
+    it(`messageStream-${i}: should deliver messages to all ${i} group members within acceptable time limits and verify stream consistency`, async () => {
       try {
         const creator = workers.getCreator();
         newGroup = (await creator.client.conversations.newGroup(
           getInboxIds(i),
         )) as Group;
-        // Use the dedicated conversation stream verification helper
-        const verifyResult = await verifyNewConversationStream(
+        await newGroup.sync();
+        await newGroup.addMembers(
+          workers.getAllButCreator().map((worker) => worker.client.inboxId),
+        );
+        await newGroup.sync();
+        const verifyResult = await verifyMessageStream(
           newGroup,
           workers.getAllButCreator(),
         );
 
-        setCustomDuration(verifyResult.averageEventTiming);
-        expect(verifyResult.allReceived).toBe(true);
-
         // Save metrics
         summaryMap[i] = {
           ...(summaryMap[i] ?? { groupSize: i }),
-          conversationStreamTimeMs: verifyResult.averageEventTiming,
+          messageStreamTimeMs: verifyResult.averageEventTiming,
         };
+
+        setCustomDuration(verifyResult.averageEventTiming);
+        expect(verifyResult.allReceived).toBe(true);
       } catch (e) {
         logError(e, expect.getState().currentTestName);
         throw e;
@@ -76,6 +79,7 @@ describe(testName, async () => {
     });
   }
 
+  // Aft
   // After all tests have run, output a concise summary of all timings per group size
   afterAll(() => {
     saveLog(summaryMap);
